backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "motor>=3.6.1",
    "bcrypt==4.0.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.4.0",
    "gunicorn>=23.0.0",
    "uvicorn[standard]>=0.33.0",
]
